[#15] restructure client app
