+++ src/web_client/index.html
@@ -5,7 +5,7 @@
     <title>ProceZeus</title>
   </head>
   <body>
-    <div id="app"></div>
+    <div id="proceZeus"></div>
     <!-- built files will be auto injected -->
   </body>
 </html>
