+++ src/web_client/src/components/Chat.vue
